+++ src/Tools/RapidResponseDesk/RapidResponseDesk.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import './CommercialData.scss';
+import './RapidResponseDesk.scss';
 import { connect } from 'react-redux';
 import { getAppropriateAuthToken } from '../../App';
 
-const CommercialData = ({ authToken }) => {
+const RapidResponseDesk = ({ authToken }) => {
   return (
     <div className="commercial-data-panel">
       <p>Commercial Data</p>
@@ -16,4 +16,4 @@ const mapStoreToProps = (store) => ({
   authToken: getAppropriateAuthToken(store.auth, store.themes.selectedThemeId),
 });
 
-export default connect(mapStoreToProps, null)(CommercialData);
+export default connect(mapStoreToProps, null)(RapidResponseDesk);
